Prevent future automatic updates after a breaking change has been detected.
Fix an issue where the Updater does not realize a breaking change has occurred between the current version and the newest version, because the "breaking change" occurred in an interim release, and not in the latest release.

i.e. 2 releases have been published since the last update, with the first one having a breaking change, and the second not having a breaking change. The Updater will skip the first release, due to the breaking change, but then will subsequently update to the second release, since it's not labeled "breaking".

In the true spirit of this script, this fix is accomplished in a "dumb" manner, by creating a flag file in the Immich directory on a breaking change and then checking for the existence of this flag file on subsequent runs. This adds a step of having to delete the file, when manual Immich updates are performed.